% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CounterMatches-class.R
\docType{class}
\name{CounterMatches-class}
\alias{CounterMatches-class}
\title{Virtual class of object that can count overall number of sequences that match
in some regions}
\description{
You can choose to use two class that inherit from this virtual class. The
class CounterMatchesMat is a matrix that contain the overall number of
sequences that match in the regions, it's easy and fast. When you have to use
less memory and you know that the matrix can be scattered, you can use
CounterMatchesList, which have a more complex internal structure and can be a
little slower, but it allows you to save space in memory.
}
\section{Slots}{

\describe{
\item{\code{arranged}}{boolean. It say whether the structure is sorted by match
number}

\item{\code{reduced}}{boolean. It say whether the sequences and the regions with no
matches are deleted from the structure}
}}

\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
