% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countseq-package.R
\docType{package}
\name{countseq-package}
\alias{countseq-package}
\alias{countseq}
\title{countseq - Overall count for each search sequence within the
interesting genomic regions.}
\description{
The \code{countseq} package implements several functions for count for each
search sequence within the interesting genomic regions.
}
\details{
\tabular{ll}{
Package: \tab countseq\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2020-04-15\cr
License: \tab GPL (>=3)\cr
}
}
\references{
\url{https://github.com/daxus4/countseq}\cr
Raffaelli (2021) Some paper. Journal, volume, issue.\cr
}
\author{
Davide Raffaelli [aut, cre]\cr
Politecnico di Milano\cr
Maintainer: Davide Raffaelli\cr
E-Mail: \href{mailto:davide2.raffaelli@mailpolimi.it}{davide2.raffaelli@mailpolimi.it}
}
