% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{countSeqs}
\alias{countSeqs}
\title{Count reads mapped on multiple regions of a genome}
\usage{
countSeqs(genome, regions, sequences)
}
\arguments{
\item{genome}{A genome of BSgenome class}

\item{regions}{A GRanges which specify the interesting regions}

\item{sequences}{A DNAStringSet which contains the sequences to be matched
within the regions}
}
\value{
Matrix of integers, the columns are the sequences, the rows are the
regions and the cells the overall counts
}
\description{
This function returns an overall count for each search sequence within the
interesting genomic regions. Also overlapping sequences.
}
\examples{

countSeqs(Hsapiens, GRanges("chr1", IRanges(1e6 + c(1,101), width=100)),
DNAStringSet(c("AA", "AT", "GG")))

}
\references{
\url{https://en.wikipedia.org/wiki/Genome}\cr
}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffarelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
