% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSeqsList.R
\name{countSeqsList}
\alias{countSeqsList}
\title{Count reads mapped on multiple regions of a genome}
\usage{
countSeqsList(genome, regions, sequences)
}
\arguments{
\item{genome}{A genome of BSgenome class}

\item{regions}{A GRanges which specify the interesting regions}

\item{sequences}{A DNAStringSet which contains the sequences to be matched
within the regions}
}
\value{
list where for each sequence that have a match there is a list of
the match with the region and the start.
}
\description{
This function returns an overall count for each search sequence within the
interesting genomic regions. Also overlapping sequences. It's the best
choice when you think that there are few sequences that match in many
regions; otherwise you an use countSeqsMat.
}
\examples{

#Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

countSeqsList(BSgenome.Hsapiens.UCSC.hg38::Hsapiens, regs, seqs)

}
\references{
\url{https://en.wikipedia.org/wiki/Genome}\cr
}
\seealso{
\code{\link{countSeqsMat}}\cr
}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
