% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CounterMatches-class.R
\name{isReduced}
\alias{isReduced}
\title{Accessory method that return if the object is reduced, that is if there are
only regions and sequences with one or more matches}
\usage{
isReduced(.object)
}
\arguments{
\item{.object}{A CounterMatches object}
}
\value{
boolean indicating if .object is reduced
}
\description{
Accessory method that return if the object is reduced, that is if there are
only regions and sequences with one or more matches
}
\examples{
#' #Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

#Get a CounterMatchesMat object
mat <- getCounterMatchesMat(BSgenome.Hsapiens.UCSC.hg38::Hsapiens, regs,
                            seqs)
#is reduced?
isReduced(mat)

}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
