% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CounterMatches-class.R
\name{matches}
\alias{matches}
\title{Accessory method that return the object with the overall number of sequences
that match in some regions}
\usage{
matches(.object)
}
\arguments{
\item{.object}{A CounterMatches object}
}
\value{
object (which class depends by class of .object) that represents the
overall number of sequences that match in some regions
}
\description{
Accessory method that return the object with the overall number of sequences
that match in some regions
}
\examples{
#' #Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

#Get a CounterMatchesMat object
mat <- getCounterMatchesMat(BSgenome.Hsapiens.UCSC.hg38::Hsapiens, regs,
                            seqs)
#Get matches
matches(mat)

}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
