% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromFile.R
\name{readFromFile}
\alias{readFromFile}
\title{Read from a file a matrix of integers, with rownames and colnames}
\usage{
readFromFile(filename)
}
\arguments{
\item{filename}{string. Name of the file where the matrix is saved}
}
\value{
matrix of integers, with rownames and colnames
}
\description{
This function read from a file a matrix obtained from
\code{\link{countSeqsMatrix}} and written into it or any other matrix which
contains only integers and with rownames and colnames
}
\examples{

#Get a matrix of integers with rownames and colnames from matrix.txt
readFromFile(system.file("extdata", "matrix.txt", package = "countseq"))

}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail:
\href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
