% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CounterMatches-class.R
\name{reduce}
\alias{reduce}
\title{Method that reduce the object}
\usage{
reduce(.object)
}
\arguments{
\item{.object}{A CounterMatches object}
}
\value{
object (same class of .object) reduced
}
\description{
This function reduce, that is delete from the object the regions and the
sequences with 0 matches
}
\examples{
#' #Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

#Get a CounterMatchesMat object
mat <- getCounterMatchesMat(BSgenome.Hsapiens.UCSC.hg38::Hsapiens, regs,
                            seqs)
#reduce mat
reduce(mat)

}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
