% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceMatrix.R
\name{reduceMatrix}
\alias{reduceMatrix}
\title{Delete rows and/or columns of matrix of numerics full of zeros}
\usage{
reduceMatrix(mat, rows = TRUE, cols = TRUE)
}
\arguments{
\item{mat}{matrix of numerics}

\item{rows}{logical. Should rows full of zero be deleted?}

\item{cols}{logical. Should columns full of zero be deleted?}
}
\value{
input matrix without rows/columns full of zeros
}
\description{
This function delete rows and/or columns of matrix of numerics full of zeros.
It is useful for \link{countseq} package, because it allows to delete regions
and sequences with no match in the matrix returned from
\code{\link{countSeqsMatrix}} function.
}
\examples{

#create matrix of numerics
mat <- matrix(c(6,4,0,0), nrow=2)

#Order matrix of matches
reduceMatrix(mat)
}
\seealso{
\code{\link{countSeqsMatrix}}\cr
}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail:
\href{mailto:davide.raffaellii@gmail.com}{davide.raffaellii@gmail.com}
}
