% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CounterMatches-class.R
\name{seqNames}
\alias{seqNames}
\title{Accessory method that return sequences' names}
\usage{
seqNames(.object)
}
\arguments{
\item{.object}{A CounterMatches object}
}
\value{
vector of characters that contains the sequences' names
}
\description{
Accessory method that return sequences' names
}
\examples{
#' #Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

#Get a CounterMatchesMat object
mat <- getCounterMatchesMat(BSgenome.Hsapiens.UCSC.hg38::Hsapiens, regs,
                            seqs)
#get sequences' name
seqNames(mat)

}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
