% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortMatrix.R
\name{sortMatrix}
\alias{sortMatrix}
\title{Sort columns of matrix of numerics}
\usage{
sortMatrix(mat, decreasing, order.rows = FALSE)
}
\arguments{
\item{mat}{matrix of numerics}

\item{decreasing}{logical. Should the sort order be decreasing?}

\item{order.rows}{logical. Should rows be ordered too?}
}
\value{
input matrix ordered depending on the parameters
}
\description{
This function sort columns of matrix of numerics. It is useful for
\link{countseq} package, because it allows to order the matrix returned from
\code{\link{countSeqsMatrix}} function.
}
\examples{

#create matrix of numerics
mat <- matrix(c(6,4,0,7), nrow=2)

#Order matrix of matches
sortMatrix(mat, FALSE, TRUE)
}
\seealso{
\code{\link{countSeqsMatrix}}\cr
}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail:
\href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
