% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CounterMatches-class.R
\name{sortSeqs}
\alias{sortSeqs}
\title{Sort the object by sequences' overall count}
\usage{
sortSeqs(.object, decreasing, order.regions)
}
\arguments{
\item{.object}{A CounterMatches object}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{order.regions}{logical. Should internal vector of counts per region
be ordered too?}
}
\value{
object (same class of .object) ordered depending on the parameters
}
\description{
This function sorts the object in decreasing or increasing order, by using as
metric the overall counts of matches for each sequence. You can also order
the regions
}
\examples{
#' #Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

#Get a CounterMatchesMat object
mat <- getCounterMatchesMat(BSgenome.Hsapiens.UCSC.hg38::Hsapiens, regs,
                            seqs)
#Sort mat
sortSeqs(mat, FALSE, TRUE)

}
\seealso{
\code{\link{sortSeqsList}}\cr
\code{\link{sortSeqsMatrix}}\cr
}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
