% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortSeqsList.R
\name{sortSeqsList}
\alias{sortSeqsList}
\title{Sort list returned from \code{\link{countSeqScattered}}}
\usage{
sortSeqsList(seqsList, decreasing, recursively)
}
\arguments{
\item{seqsList}{A list returned from countSeqScattered}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{recursively}{logical. Should internal vector of counts per region
be ordered too?}
}
\value{
input list ordered depending on the parameters
}
\description{
This function sort list returned from countSeqScattered in decreasing or
increasing order, by using as metric the overall counts of matches for each
sequence. To order the matrix returned from countSeq please use
sortSeqsMatrix
}
\examples{

#Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

#Get the list of matches
listSeqs <- countSeqsScattered(BSgenome.Hsapiens.UCSC.hg38::Hsapiens,
 regs, seqs)

#Order list of matches
sortSeqsList(listSeqs, FALSE, TRUE)

}
\seealso{
\code{\link{countSeqScattered}}\cr
\code{\link{sortSeqsMatrix}}\cr
}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
