% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortSeqsMatrix.R
\name{sortSeqsMat}
\alias{sortSeqsMat}
\title{Sort matrix returned from countSeqMat}
\usage{
sortSeqsMat(seqsMat, decreasing, order.regions)
}
\arguments{
\item{seqsMat}{A matrix returned from countSeqMat}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{order.regions}{logical. Should regions be ordered too?}
}
\value{
input matrix ordered depending on the parameters
}
\description{
This function sort matrix returned from countSeqMat in decreasing or
increasing order, by using as metric the overall counts of matches for each
sequence. To order the list returned from countSeqList please use
sortSeqsList
}
\examples{

#Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

#Get the matrix of matches
matSeqs <- countSeqsMat(BSgenome.Hsapiens.UCSC.hg38::Hsapiens,
  regs, seqs)

#Order matrix of matches
sortSeqsMat(matSeqs, FALSE, TRUE)

}
\seealso{
\code{\link{countSeqMat}}\cr
\code{\link{sortSeqsList}}\cr
}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
