% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToFile.R
\name{writeToFile}
\alias{writeToFile}
\title{Write to a file a matrix obtained from \code{\link{countSeqsMatrix}}}
\usage{
writeToFile(matrix, filename)
}
\arguments{
\item{matrix}{A matrix returned from \code{\link{countSeqsMatrix}} or any
matrix which contains only integers and with rownames and colnames}

\item{filename}{string. Name of the file where the matrix will be saved}
}
\value{
TRUE if the file is created, FALSE otherwise
}
\description{
This function write to a file a matrix obtained from
\code{\link{countSeqsMatrix}} or any other matrix which contains only
integers and with rownames and colnames
}
\examples{

#Get a matrix of integers with rownames and colnames
mat <- matrix(data = c(1,2,3,4), ncol = 2)
rownames(mat) <- c('1','2')
colnames(mat) <- c('1','2')

#Write mat to a file
writeToFile(mat, "matrix.txt")

}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail:
\href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
