% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countseq-package.R
\docType{package}
\name{countseq-package}
\alias{countseq-package}
\alias{countseq}
\title{countseq - Overall count for each search sequence within the
interesting genomic regions.}
\description{
The \code{countseq} package implements several functions for count for each
search sequence within the interesting genomic regions. It contains methods
to save and load results in .txt files.
This package is made for a university project.
}
\details{
\tabular{ll}{
Package: \tab countseq\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2020-04-15\cr
License: \tab GPL (>=3)\cr
}
}
\references{
\url{https://github.com/daxus4/countseq}\cr
Raffaelli (2021) Politecnico di Milano\cr
}
\author{
Davide Raffaelli [aut, cre]\cr
Politecnico di Milano\cr
Maintainer: Davide Raffaelli\cr
E-Mail: \cr
\href{mailto:davide2.raffaelli@mailpolimi.it}{davide2.raffaelli@mailpolimi.it}
}
