% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSeqsMatrix.R
\name{countSeqsMatrix}
\alias{countSeqsMatrix}
\title{Count reads mapped on multiple regions of a genome}
\usage{
countSeqsMatrixrix(genome, regions, sequences, reduced = TRUE)
}
\arguments{
\item{genome}{A genome of BSgenome class}

\item{regions}{A GRanges which specify the interesting regions}

\item{sequences}{A DNAStringSet which contains the sequences to be matched
within the regions}

\item{reduced}{boolean. It specify if the columns and rows full of zeros have
to be deleted}
}
\value{
Matrix of integers, the columns are the sequences, the rows are the
regions and the cells the overall counts
}
\description{
This function returns an overall count for each search sequence within the
interesting genomic regions. In the count are counted also overlapping
sequences.
}
\examples{

#Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

countSeqsMatrix(BSgenome.Hsapiens.UCSC.hg38::Hsapiens, regs, seqs)

}
\references{
\url{https://en.wikipedia.org/wiki/Genome}\cr
}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail: \href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
