% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSeqsMatrix.R
\name{countSeqsMatrix}
\alias{countSeqsMatrix}
\title{Count reads mapped on multiple regions of a genome}
\usage{
countSeqsMatrix(genome, regions, sequences, reduced = FALSE,
ordered = FALSE)
}
\arguments{
\item{genome}{A genome which inherits from \link{BSgenome} class}

\item{regions}{A \link{GRanges} which specify the interesting regions}

\item{sequences}{A \link{DNAStringSet} or a vector of characters which
contains the sequences to be matched within the regions}

\item{reduced}{boolean. It specify if the columns and rows full of zeros have
to be deleted. If you want a different reduce options please use
\code{\link{reduceMatrix}}}

\item{ordered}{boolean. It specify if the sequences should be ordered in
descreasing order. If you want a different order options please use
\code{\link{sortMatrix}}}
}
\value{
Matrix of integers, the columns are the sequences, the rows are the
regions and the cells the overall counts of matches
}
\description{
This function returns an overall count for each search sequence within the
interesting genomic regions. In the count are counted also overlapping
sequences.
}
\examples{

#Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

countSeqsMatrix(BSgenome.Hsapiens.UCSC.hg38::Hsapiens, regs, seqs)

}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail:
\href{mailto:davide.raffaellii@gmail.com}{davide.raffaellii@gmail.com}
}
