% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortSeqsMatrix.R
\name{sortSeqsMatrix}
\alias{sortSeqsMatrix}
\title{Sort matrix returned from countSeqsMatrix}
\usage{
sortSeqsMatrix(seqsMat, decreasing, order.regions = FALSE)
}
\arguments{
\item{seqsMat}{A matrix returned from countSeqsMatrix}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{order.regions}{logical. Should regions be ordered too?}
}
\value{
input matrix ordered depending on the parameters
}
\description{
This function sort matrix returned from countSeqsMatrix in decreasing or
increasing order, by using as metric the overall counts of matches for each
sequence.
}
\examples{

#Charge Hsapiens genome
library(BSgenome.Hsapiens.UCSC.hg38)

#Get a GRanges object which contains regions of interest
library(GenomicRanges)
regs <-GRanges("chr1", IRanges(1e6 + c(1,101), width=100))

#Get a DNAStringSet which contains the sequences to be mapped
library(Biostrings)
seqs <- DNAStringSet(c("AA", "AT", "GG"))

#Get the matrix of matches
matSeqs <- countSeqsMatrix(BSgenome.Hsapiens.UCSC.hg38::Hsapiens,
                            regs, seqs)

#Order matrix of matches
sortSeqsMatrix(matSeqs, FALSE, TRUE)

}
\seealso{
\code{\link{countSeqsMatrix}}\cr
}
\author{
Davide Raffaelli\cr Politecnico di Milano\cr Maintainer: Davide
Raffaelli\cr E-Mail:
\href{mailto:davide2.raffaelli@mail.polimi.it}{davide2.raffaelli@mail.polimi.it}
}
